Move all files to project root
